requirements.txt
playwright>=1.50.0
pytest>=7.0.0
pytest-html>=4.0.0
pytest-xdist>=3.0.0
pytest-playwright>=0.4.0
requests>=2.25.0
httpx>=0.24.0
python-dotenv>=0.19.0
pydantic>=2.0.0
